Revert^2 "Better detect PrimitiveArrayCritical issues"
This reverts commit 3ba842dfa15bb51d6579b0ac4dc92f6fbc23db23.

Reason for revert: Re-land fixed version of original CL

PS1 is identical to aosp/2089666 .

PS2 performs the checks using the raw disable_thread_flip_count field
instead of the getter, so that it also works with the read barrier
disabled. In that case this field should always be zero, but it
doesn't hurt to check.

Bug: 231038541
Test: Build and boot AOSP

Change-Id: I78165d5fa8e82a8aed5a61da3d9a6a8ec8bfa69e
Signed-off-by: minaripenguin37 <alexfinhart@gmail.com>